Use default password for external console access
This adds the default password back to the Karaf console by default on
uninteractive install when we enable external interfaces for the
console. This has been done to keep some measure of the security of the
console intact. This will also prevent confusion from users who expect
the default password to be needed when logging in. See the following
page for more details: <https://community.openhab.org/t/124922>.

Signed-off-by: Ethan Dye <mrtops03@gmail.com>