Automatically fix `.bash_profile`
This is necessary to fix the issue caused by 253dd15 which caused the
functions to check if openHAB is installed to no longer be located where
the installed version of `.bash_profile` expected them to be. The
`.bash_profile` included in the project was fixed for this issue in
commit c34eccc and this ensures that the issue is fixed in all
installations once openhabian-config is run.

Signed-off-by: Ethan Dye <mrtops03@gmail.com>